Add coreos.configdrive=0 as a default setting
Due to a race condition in CoreOS, any pre-existing configuration
drive that is present on a system is being attached to the node.

The only documented way to prevent CoreOS from attaching a
pre-existing configuration drive is to pass the kernel parameter.

This combined with an update that has been made to IPA, closes #2.

Additional details can be found at
https://bugs.launchpad.net/ironic/+bug/1433812