Add playbook to remove deployment images from the system
In the event of redeployments on installations that have been
in-place for some time, it may be necessary to clean the
master_images folder along with the deployment kernel and
ramdisk to ensure that fresh images are downloaded during
the next installation.

Updated troubleshooting documentation to mention availability of
the playbook and the potential case when it may be needed.