Additional minor cleanup
Changed a tar command to the ansible unarchive module and removed
local_action from a copy command that appears in ansible-lint
reporting.

As for the copy command, The copy command it's self does not work
with our specific use case, so we will continue to use command to
directly invoke cp.