Changes to support CI testing
When we begin using CI testing, we will want to support installation
of specific revisions which ultimately means Bifrost's installation
process needs to support a slightly different path.

Adds a setting, ci_testing, which is normally set to false.

This setting skips retreval of the files for Ironic, ironicclient,
and shade so any CI prepatory can just leave the appropriate files
in /opt/stack/.