Script improvements for greater compatability/testing
Changed the scripts to support yum as well as apt, and added a
step to execute the ansible check-syntax sequence in order to fail
fast in the event of a syntax issue.

Additionally, the installation of jinja2 has been switched to use
a requirements.txt file placed in the root of the repository.