Change to Ansible core modules devel branch
Switching to the devel branch until the submodule
pointer gets updated so bifrost is operational again.

This was due to the Ansible mysql_user module breaking when
no configuration file was explicitly set.  A fix has been landed
however the pointer for the main Ansible repo naturally lags
behind the current development state of the modules repo.