Use Cirros for CI testing
Utilizing Cirros for the guest operating system when performing a
basic test saves a significant amount of time and resources for the
test as it prevents the creation of a disk image.

Additionally, the native support for configuration drives allows
the amount of time we're waiting on the guest OS to be greatly
reduced.

Change-Id: I5a4c497d3502c975db060d3ee08ea2056399a062
Closes-Bug: #1460052