Refactor pip check/install & add dependency for CentOS
Use easy_install to install pip to avoid additional dependencies
on CentOS. Refactor the check for an installed pip since it will
now be the same in both OS families.

env-setup.sh was missing a package required if the user is
starting from a very basic CentOS install.

Also fix a minor typo in the script.

Change-Id: I1ee783862b2bc5ad8b5df442ed7ca25732d8f706