Refactor ironic-install playbook to support CentOS & RHEL
Refactor the ironic-install role so that it can be used on
Centos and RHEL. Also add a new template for systemd configuration
so we can install and run the ironic services.

Document the additional steps that RHEL and CentOS users will
need to take to get bifrost up and running.

Install gcc in env-setup.sh as it is needed to install pip and
we can't take it for granted.

Temporarily remove amttool from RedHat family dependencies since
it's not currently packaged for RHEL or CentOS 7. We'll handle this
in a separate patch.

Change libxslt-dev to libxslt1-dev as it is the correct name for the
package.

Closes-Bug: 1460930
Closes-Bug: 1456896
Partial-Bug: 1463871

Change-Id: I72a5b329d3c55350644a777b5a77d61947c921bc