Move explicit URLs in ironic-install to defaults
To facilitate offline installs, any explicit URLs should be
moved to a location where they can be easily overridden with
a local equivalent.

That way, a default install will use the canonical location, but
a user can easily override this behavior in inventory.

(Such gymnastics are not necessary for apt/yum/pip, because those
can be configured to point to local/alternate repositories independently).

** WIP - Needs *better* documentation **

Change-Id: I9dada2fe769e30abc06dac092b9f47b99969a954
Closes-Bug: #1464787