Attempt to upgrade iPXE if no mac found
Apparently older iPXE implementations that may be burned into
ROMs may not recognize the template MAC syntax, and thus be unable to
boot.

In the event of this happening, we should attempt to chainload the
iPXE image.

Additionally removed autofree as this may not be recognized by
some versions of iPXE.

Change-Id: I6e458d6a9383fb2d31d743b0124384a2036520df