Disambiguate the roles of ci_testing, ci_testing_zuul etc
We currently use both ci_testing and ci_testing_zuul to define
whether we are being run from CI. However, ci_testing_zuul
required a hostname match to be defined, and ci_testing
depended on the ZUUL_CHANGES environment variable, set in
OpenStack CI. This was a little confusing.

This change splits up the boolean values which control
CI-related tasks in the following way:

ci_testing - Used for general-purpose CI tasks that are not
tied to the OpenStack CI infrastructure
ci_testing_zuul - Used for CI tasks that are specific to
the OpenStack CI infrastructure
ci_testing_zuul_changes - Used for CI tasks when
ZUUL_CHANGES is set.

Also, move away from the old hostname-based check to confirm
whether we are running under OpenStack CI as this could prove
to be fragile. Instead, look for the ZUUL_BRANCH environment
variable.

Change-Id: Ica8e84886c85911676c711ddcdd61604de8f3ee4