Add logic to allow bifrost_inventory to run without shade
The dynamic inventory script only requires shade when querying
ironic for a list of hardware resources.  As we may not have
completed the installation sequence, shade may not be present.

Added a check to attempt to load shade and then error if a user
attempts to use the script to query ironic with shade missing.

Change-Id: I300e043bd76d4872781b42e05a5eb76a30a74f5e