Allow env-setup.sh to use alternate url/branches
For an offline install, env-setup.sh needs to be able to clone
from locations other than GitHub.  Additionally, since
the ansible repo will have to have its .gitmodules file patched
to point submodules to some other location as well, its convenient
to allow for alternate branches.

As a side-effect, this moves the documentation of what branch
we're using to the top of the script for easy reference and removes
the duplicated branch name.

Change-Id: I7432531915c03638aaf52f836cda607d424351ba