Move diskimage-builder and bootable image creation to roles
Break out image creation via diskimage-builder into a separate role.
Also break out create-bootable-image for consistency, though it is
legacy stuff that we do not expect to use in future.

In addition to the above, clean up a couple of typos, etc. in the
adjacent code and documentation.

Implements: blueprint bifrost-deployment-image
Change-Id: If3dbfa6a6b4e3f6eee9b614342dbf5bccadfbd65