Switch to simple-init for network configuration
Bifrost defining and attempting to bake in configuration files
into configuration drives to allow initial configuration of
connectivity is a crazy idea.

Since the Openstack-infra folks have glean and simple-init, we
will utilize their tooling to provide greater flexibility across
operating systems.

Additionally, the ability to write out a simple debian style
interfaces config file has been retained, however off by default
as it is necessary for cirros testing.  In addition to this, the
ability for inventory based data to override the contents of the
network_info.json file.

Change-Id: Id85b18277b4d78acf418f63f046e39e889590719
Closes-Bug: 1463191