Removal of legacy roles and related updates
Removed legacy roles that utilize the baremetal.csv and looping
to achieve the desired results which was extremely limiting and the
predicessor to the dynamic inventory module and roles which
allow greater flexibility.

Additionally updated the deploy and enroll yaml files to redirect
users to the newer process and updated the README to remove
references to the older process.

Change-Id: Ia394652686608c012a848c0329195fba9abd45d1