Change ironic configuration to utilize sample file
Changed the install-ironic role to utilize the sample ironic.conf
file supplied with ironic for the new installation, which leverages
Ansible lineinfile and sed to modify the file to be in the required
state.

By and large, this change will address the bug the most, since
ironic's configuration file is the file we want to track, however
other configuration template files need to be reviewed.

Change-Id: I45c4543c069805a3e0c9989c63852712bacbefb5
Partial-Bug: 1471985