Clarify variable names & cleanup docs
The deploy_kernel* and deploy_ramdisk* variables were confusingly
named because they looked similar to deploy_image. I've renamed
them to ipa_kernel* and ipa_ramdisk* to reflect that they are the
CoreOS kernel and ramdisk from IPA and have updated the
documentation. I also cleaned up a minor issue where we defined
a path for the local copies of the kernel and ramdisk but didn't
use it in the install playbook.

I've also removed deploy_image_upstream_url from the offline
installation documentation as we no longer use it.

Partial-Bug: 1478726
Change-Id: Ia8c865521974e6bdd9e8e6a3ffa4c5f6b1f2f6ad