Update default OS and support dib_packages
In order to utilize ubuntu-minimal, we need to install an additional
package in order to login to the host.  As such, we have updated the
playbooks to support the dib_packages variable to allow a user to
explicitly state if they want to install openssh-server, or any
other required software when building their image.

At this time, staying with debian as the default image, although
it does ship with cloud-init, it does not interfere with local
testing.

Change-Id: Ic537d2b1e13cc35379d77f078bcc1a1faeb3f526