Kludge to make Fedora 22 package installs work
Fedora 22 uses dnf, but the ansible release we use sets the
ansible_pkg_mgr fact to yum, which won't work if the relevant
python libraries aren't installed. Reset the fact in playbooks
which install packages to fix package installation for now.
(This change can be safely removed when we start using an ansible
release which prefers dnf.)

Change-Id: If2dd0d45de84a98e627f322a5579b2482e0204c3