Minor testing clean-up
After changing from ubuntu to debian, testing was never
updated to validate that the hostname was not ubuntu.

Since we support simple-init, we should check for multiple
default OS names and fail accordingly if they are found.

Additionally, changed the default testing user to root
as simple init writes the key to the root user's
authorized_keys file.

Change-Id: Icce62d0b9256188dbc9b49d948c99bb0c469ae6b