Move DIB components to support depends-on
Until support for zuul-cloner arrives for bifrost's CI settings,
we need mechanisms to allow us test against a dependent revision
in the diskimage-builder repositioy.

As such, we've moved the diskimage-builder download and installation
to align with the other packages from source, and updated the CI
prep role to apply dependent revisions.

Additionally, correct conditionals in the OpenStack CI prep role.

Change-Id: Ie5db11ff7d2160c7cea9a71aebd86d1b369983c1