Rename install role to bifrost-ironic-install
Since the install role is specifically geared to install ironic
in a way that bifrost uses, we should name the role appropriately.

Note: This will be followed up with a change to revise the
README file for the role as changing and renaming the file
will result in issues with the file history.

Change-Id: I553a29206a045b5aa0e679ce05e984c7dd4b9df3
Depends-On: Ia6226a8573753ff02da19dcb3c2b5abfb9b3c297