Remove the group_vars/all file for easy customization
Removal and split of the group_vars all file resulted in a general
clean-up and optomization of dependency information with-in the
roles as defaults are required and a number of them were being
supplied from group_vars/all.

Additionally split out the group_vars/all file into
group_vars/baremetal and group_vars/localhost files with
improved comments to help direct a user wanting to perform
deeper customization.

Due to variable loading processes, additionally unwire the bifrost-
create-bootable-image role, as it is no longer developed upon and
the default vars are conflicting with the variables used across
role.

Additionally updated READMEs where appropriate.

Co-Authored-By: stephane <stephane@alum.mit.edu>
Change-Id: Ia6226a8573753ff02da19dcb3c2b5abfb9b3c297