Support for fine-grained distro-related defaults
Currently, we specify defaults for the RedHat and Ubuntu
distribution families, and perform adjustments to these
in the playbooks based on checks for OS version, etc.
This is awkward. Instead, import defaults files in increasing
order of specificity to tune the install for each distribution
and version without extra playbook items.

Change-Id: I60b5a1f1c1f63fe4c742599a5a8a12a80cd15652