Add static DHCP for dynamic inventory
In some cases it is simpler to just configure the images being deployed
to use DHCP, rather than teach them how to put the IP address in the
configdrive in the right place. This will setup dnsmasq to send the
static address for the box. It will use that address for the deploy as
well.

Co-Authored-By: Clint Byrum <clint@fewbar.com>
Co-Authored-By: Gregory Haynes <greg@greghaynes.net>
Co-Authored-By: Julia Kreger <juliaashleykreger@gmail.com>
Change-Id: I524958be5e787d42c91278baf2c4a14beb965e7c