Update iPXE agent template and boot.ipxe ownership
At some point the Ironic iPXE configuration changed.
The params now passed are full URLs and it is no longer
necessary for us to write the beginning of the URL out.

Additionally update the ownership of the boot.ipxe file so
ironic is able to overwrite it.  Far from ideal, but required
at this time.

Closes-Bug: #1490546
Closes-Bug: #1490557
Change-Id: I1a54ac6dce1497823795f896907663b44ebaa67d