Correct pxe template for CoreOS config drive
In a previous commit, we erroniously removed the setting to disable
CoreOS from loading configuration drive files.  Turns out that the
solution in Ironic was to update the template files, and to await
for CoreOS to remedy the issue, which has not occured.

This ultimately results in every other deployment to a machine
to be missing a configuration drive since we removed the setting.

As such, we need to put the setting back, this time in the template
file we utilize.

Previous commit: I0e044121e146f3036aa6b1830983dd9d7a68e9ac

Change-Id: I7d4a9c165f069865a26f6900f64046485f00cdee