CORS configuration support
Since CORS support has landed in Ironic, we should enable a bifrost
user to be able to turn CORS on.

Note: Presently the ironic-webclient sends an x-client header which
causes the oslo middlewear portion that supplies CORS to reject the
request.  A change is pending to remove that from the webclient,
however if a user wishes to perform testing with the webclient,
x-client will need to be added to the list of allowed headers.

Depends-On: I23e902c8637e142fba23d71467225d48ee265253
Change-Id: I487a57c54995e0572072aa54193e985ea257ed70