Move dnsmasq to the end of the Debian package list
On Ubuntu, apt automatically starts dnsmasq upon installing it. This
temporarily causes a blip in the system's ability to resolve domain
names. In a bifrost install, this becomes apparent when the next
package ansible attempts to install, rabbitmq-server, fails with the
error "Could not resolve 'archive.ubuntu.com'", while subsequent
packages and subsequent runs of the bifrost install step succeed. This
patch moves the dnsmasq to the end of the required package list so that
its installation does not interfere with the installation of the rest
of the required packages.

An alternative to this fix could be to configure policy-rc.d to not
start services by default. This would require making sure all services
are managed in bifrost-ironic-install/tasks/main.yml and would be a bit
more dramatic of a change.

Change-Id: If1558d76ce3f652c0798b459678baea9f8b26fb0