Fix default log file location
The default log file location was not remedied by change
I76a2a23e2a75022bae3511700c81145b5cbeae77.  This fix changes
the LOG_LOCATION such that the logs folder is always defined
instead of using the workspace root which is not compatible
with the test jobs presently defined for bifrost.

Depends-On: I23e902c8637e142fba23d71467225d48ee265253
Change-Id: I9329d9485c7d5ab3d51831446993df47b70d881e
Closes-Bug: 1490532