Install ironicclient after installing the shade library
Shade needs to be installed before ironicclient, as ironicclient is a
dependency, pip ends up removing the previously installed version when
we force the upgrade operation, which may downgrade the library to the
released version which may be undesirable.

Additionally set the pip install of ironicclient to be latest instead
of being present.

Depends-On: I23e902c8637e142fba23d71467225d48ee265253
Change-Id: I3b1dca09e95e7a29b9f7e4b0a048aaa46bb2e22e