Removed integrated ZUUL reference patcher
Bifrost intiially had support for parsing ZUUL_CHANGES, however
due to changes in it's gate CI jobs, these steps are no longer
necessary.

As a result of this change, we also need to support copying
files directly and not forcibly resetting the git status since
the repositories are provided to us in the way they are needed
without additional work.  As a result, added logic to the
preparation and git download role to support local file copies
instead of git downloads.

Also changed default source location for CI testing repos to
leverage the WORKSPACE environment variable.

Change-Id: I23e902c8637e142fba23d71467225d48ee265253
Depends-On: I01c16990dc66988fab29c41ec4b4e6094213b2ca