Initial support for installation of ironic-inspector
This review adds the ironic-inspector as a component that can be
opted to be installed by the user.  Note, this is moderately useless
without shade support and modules to assist the users to leverage
this workflow.

If a user wishes to utilize at this time before those items exist
then they will have to move nodes in available state back to
manageable state, and then invoke the inspect state which will
return them to manageable state once done.

Change-Id: I7b4ff92fa27578a9a7b0f25fc6e8658c3f2700aa
implements: blueprint bifrost-inspector-support
Depends-On: I27caa1122a72ac655958b7a6aa14b7566964f998