Stop using distro setuptools and easy_install
distro setuptools is old and should never be touched if you're going
to install things from pip. easy_install is insecure and open to
MITM attacks.

Downloading get-pip.py over https and running it will install modern
pip and setuptools on the system and is the preferred method for
installing pip in all cases.

Change-Id: I1bf88f018806142ca639f3c183598b318636e61d