Fix boot.ipxe config behavior to ensure proper use
By default, on each deploy, ironic will replace the boot.ipxe file,
that a node will normally utilize to boot with the base template from
ironic. While this is good for the largely generic use case, bifrost
users have encountered various issues utilizing iPXE with different
components that we've incorporated support for.

As such, we will set the ironic config so that our template is
utilized.  Additionally updated the role README to inform and
direct users of this to file bugs if issues are encountered.

Change-Id: I8f47f4573dcf2346b4c5d424df5aabe62ebafa94
Closes-Bug: # 1504533