Correct conditionals and dnsmasq reload for deploy
The conditionals in the deployment role were incorrect if the
inventory data source was set to ironic where the interpretted
reply from the API could include an empty instance_info dictonary.

Changed the conditionals to handle the condition where the
dictonary may be empty.

Additionally changed the dnsmasq handling to directly HUP the
dnsmasq process as the service init file, if present, may not
support reloaded state as Ansible passes it directly through.

Change-Id: Id84b8869ef5fc9deb1eee8cc304ee00713a3016a