Modularize inventory and add initial testing
The inventory module for bifrost is a vital component, however
as it was developed as a single file and it needs to to be broken
up into a modular chunk in order to have meaningful testing.

Broke into module, symlinking the file into place for now until
documentation and install process can be updated.

Added two basic tests to lay the groundwork for basic testing
of the data parsing and result passing.

Change-Id: I02e79b635202a7933ae641b9d27cc19a75b7bc4f
Partial-Bug: 1499801