Add functional testing of CSV file parsing
The CSV file parsing feature of bifrost is ultimately the most logic
heavy feature of the inventory and thus sensitive to change.

As a result, we need test supplying input and see if our expected
output results. Variations added including base ssh/vm testing,
ipmi with and without bridging, and if a DHCP address is defined.
Testing the results of these combonations should help identify any
changes to the CSV parsing that may break users.

As part of this, it was realized the CSV example lacks a driver filed
example, and minor tweak of the unit testing file.

Change-Id: Id50db2387dac32a9c16e927db309a32a7806d7cc
Partial-Bug: 1499801