Remove dependency on instance_info for DHCP config
According to the documentation[1], instance_info variable is used to
define the instance that will be deployed, and has nothing to do with
its networking configuration. It therefore is not sensible to only
allow static DHCP settings for nodes with instance_info defined since
"By default, this is NOT expected to be defined".

Moreover, when the new dhcp-hosts file is created in
bifrost.dhcp-hosts.d/, dnsmasq needs to re-read this directory
regardless of whether instance_info is defined, so it does not make
sense for the "Sending dnsmasq HUP" task to be dependent on
instance_info NOT being defined.

Sending HUP should be safe to do; according to the dnsmasq man page,

  When it receives a SIGHUP, dnsmasq clears its cache and then
  re-loads /etc/hosts and /etc/ethers and any file given by
  --dhcp-hostsfile, --dhcp-hostsdir, --dhcp-optsfile, --dhcp-optsdir,
  --addn-hosts or --hostsdir. The dhcp lease change script is called
  for all existing DHCP leases. If --no-poll is set SIGHUP also
  re-reads /etc/resolv.conf. SIGHUP does NOT re-read the configuration
  file."

This does not imply any disruption of service. This assumption is
consistent with a reading of the dnsmasq signal handling code[2].

[1] http://git.openstack.org/cgit/openstack/bifrost/tree/playbooks/roles/bifrost-deploy-nodes-dynamic/README.md#n45
[2] http://bazaar.launchpad.net/~vcs-imports/dnsmasq/master/view/head:/src/dnsmasq.c#L1193

Change-Id: Iae0dae8ad4dd7c9f2263b97875702ed0970761cf