Provide dhcp-range even for static inventory
The dnsmasq man page says, regarding the dhcp-host option:

  Addresses allocated like this are not constrained to be in the range
  given by the --dhcp-range option, but they must be in the same subnet
  as some valid dhcp-range.

This confusingly-worded datum means that even when dhcp-host is
defined, which we do when setting inventory_dhcp=true, we need to
specify some kind of dhcp-range. There are a few valid arguments we
could supply to the dhcp-range parameter, but the easiest is just to go
back to using the existing dhcp_pool_{start,end} parameters, from which
dnsmasq will infer a subnet.

An alternative to this is to be more strict by using the 'static'
keyword for dhcp-range. We would still need to specify the
dhcp_pool_start value, which is a required argument for dhcp-range.

Change-Id: I4d09c3a716b864c0b0d74cc83a2f0ac7414b386c