Change default disk format for default install
The default image created during the bifrost-ironic-install phase is a
qcow2 image, not a raw image. The instance_info for the default install
type should reflect that.

The consequence of providing "raw" as the image type when the image is
not a raw image, along with keeping the default value of the new
stream_raw_images parameter[1], is that the ironic-python-agent is
not able to correctly write the image, which leads to a failure to
write the configdrive[2].

[1] http://git.openstack.org/cgit/openstack/ironic/commit/ironic?id=ce3878176e20d14681d11b53f4fc5cf1cb63e956
[2] http://paste.openstack.org/show/480078/

Change-Id: I5e948672a8af967b7eeebda45ed6251286145ff1