Allow user to set DIB_RELEASE for disk images
This change has two parts.  First it allows a user to define an
argument to the role that is wraped around diskimage-builder to
add an optional environment variable of DIB_RELEASE to the build
process.

Secondly it sets a default for the test sequence to utilize debian
jessie for the builds to stabilize the non-voting CI test.

Change-Id: Ifbc0f1ce42f3ab9ec73951a813ab0c4a38639550
Closes-Bug: #1521313