Send extra_kernel_options on pxe_append_params
When we started using ironic pxe template, we lost the ability
to send extra_kernel_options. According to our conversations, this
should be sent on pxe_append_params.
As this element was already being used for screen, I decided to make
this step a bit more generalist, and also send the extra_kernel_options
on same task.

Change-Id: I6c4a5eeb64185cc91c29b169349fff426cc53334