Add VM creation support on Debian jessie
Debian jessie changed the libvirt service name from the prior
release and as such we need additional logic to load the defaults
and a defaults file that represents jessie.

Additionally, we need to ensure that we start the default network
as it may not be started by default on Debian jessie.

Change-Id: I48a64f61d8ba42f76c2df3b984d3090b68e3cf93
Partial-Bug: 1522552