Fix dhcp range for static inventory
For static inventory, we need to define dhcp range, but
we need to provide the static flag. If we don't add the flag, and
there are more NICs attached to that network, dnsmasq will continue
providing dynamic ips on that range, causing the nodes expecting
this one to fail because the ip is already assigned.
Also, set the lease time to infinite, as we have static assignment
on that case, so no need to renew.

Change-Id: I6c0e5a3974900a4e71608e7e6c2c3fcfd1c69463