Adjust primary bifrost settings for Debian jessie support
Specific settings are required for Debian jessie as it uses
different defaults from wheezy.

Additionally, we're going to attempt to ensure that the service
folder exists if we're using a systemd template upon install
as sometimes that folder may not be present on systems.

Added genisoimage as a fallback to mkisofs in configuration
drive generation.

Change-Id: I74bf93416513094293bc071fd40e62f4240d61bc
Closes-Bug: 1522552