Add JSON/YAML inventory consumption tests
Bifrost's inventory parsing code supports parsing json and yaml
data input.

As such, we need to test that we are producing the expected
output from passing data into the general data parser,
including re-parsing the output of a previous conversion
to ensure that it is unchanged so logic parity remains
across conversion methods for any new items added.
Closes-Bug: #1527380
Change-Id: I95b3941f898b14bffa3a4a6fc6972184889cfedc