Modify detection of iso_gen_utility
On some platforms, mkisofs is a symlink to genisoimage.
Currently, the playbook will see this as mkisofs existing
on the system, so the playbook's output will look wrong
even though the correct tool is getting used. Make the output
clearer by enhancing the check to set the iso_gen_utility
fact.

Change-Id: If2c3e22c90ea30e66bd6c32e748de66548ff5d0d