Ansible 2.0 change and module cleanup
Added a new release note highlighting the change to Ansible 2.0,
as well as removal of modueles as they were included with the
release of Ansible 2.0.

Additionally added a mechanism to pull the modules from the Ansible
stable-2.0 branch should they be missing, largely in event of an
ansible 1.9.x user.

Change-Id: I0abd9bacfab020fdc7b94bb086375c27e4987693