Add initial support for EFI booting
In order to support EFI booting, we need to respond to the DHCP request
with the appropriate binary for booting the node.

At present, holding off on documentation until this has had an
opportunity to be fully tested and that we're sure that EFI booting
works without issues.

Change-Id: Ibdbeedb481796da03351daf56c91b8bd50480470
Co-Authored-By: Pierre Riteau <pierre@stackhpc.com>
Story: #1524845
Task: #8755