Correct inventory script key usage
A user reported that they were unable to utilize an inventory that
did not have the ipv4_address key present.

The user, in the bug report indicated a patch, however the end
for bifrost appears to be slightly different since the key could
exist but the value could still be None.

Added unit test to catch the minimal variable case moving forward.
Closes-Bug: #1535117
Change-Id: I0e685342887eaba054a8f80eb529835e48ec6744