Stop using pip 8.0
Bifrost automatically installed the latest pip, primarily for cases
where pip was not present.  Instead of automatically doing this each
time, we will utilize the pre-existing pip package, and if it is
unavailable, we will install an appropriate version.

Changes in the requiremnets file, beyond removal of pip, are to
facilitate the re-alignment with global requirements as the changed
entry in question pulled in one of the packages that initially
caused issues for pip 8.

Change-Id: Ic01dd9946f721ca9ad90ad87be47ab0064078301
Closes-Bug: #1536627