Rename of the nginx_port variable
The nginx_port variable is insufficiently generic enough in the event
a user really wanted to run Apache, or some other web server.

As such, this commit renames the setting to file_url_port to
prevent any confusion meanwhile supporting the ability to continue to
use an override setting should the variable nginx_port be defined in
an externally generated variable setting file.

Change-Id: Ida88cdec6ba162920df09dbdd335a9ad7868c842