Add script and playbook to test dhcp
It will run a play where it will spin up five vms, and will
truncate inventory to consider only three. The play will
configure dns using dynamic inventory, and will take an
extra parameter to specify if we want static ips or not.
It will perform a test on dnsmasq leases file, to ensure that
the dhcp behaviour is working as expected, depending on
the settings.

Change-Id: If56488f78524ceba201b50f76df306620744cda5