Add option to disabled default gateway on bifrost
There are situations where we can have different nics
living together, and the default gateway needs to be set
externally, not in the nic managed by bifrost.
The way dnsmasq works, is that if you do not set any
ip for default gateway, it assumes that to be the dnsmasq management
ip, and creates a default gateway in all nics attached to it.
That prevents default gateway being set externally, causing errors
of duplicated gateways.
To disable it, we need to pass a simple dhcp-option=3. So adding
this possibility to the template, just allowing a dnsmasq_router=false
to be passed as option.

Change-Id: I6279242de9bedf35d95c048fc2ce586435c3850b