Correct cleaning configuration
Change I5b4fbac1753cb4213eb151b663490f381b751597 landed which
changed the default cleaning setting to automated_clean from
clean_nodes. As a result of the lineinfile behavior mixed with
the configuration in the yaml, the configuration was being set
at the bottom of the file instead of at the intended location.

Added the new setting, noted the old setting is to be removed in
Newton, and hopefully corrected the syntax such that the
bifrost gate will be functional again.

Change-Id: Ieaf378e5b8b887c82f23cae9e610666dc285e872
Closes-Bug: #1548341