Make install-ironic skip IPA download if create_ipa_image == true
The flag "create_ipa_image: true" may be ignored if the IPA image file
is already present on the local disk.

However, because "download_ipa" defaults to true and runs first, the IPA
image file is always present before the tests to determine whether to
create it or not (unless one specifically sets "download_ipa: false").

This change causes the "create_ipa_image: true" flag to override
the behavior of the "download_ipa" flag.

Change-Id: I5f8e8f1d8893a279aae5b733cde5c07265197f5d