DHCP test improvements
The DHCP test ultimately relies upon bifrost putting in place
the correct DHCP configuration to reliably boot machines, and then
(hopefully) ensure that no unknown machines are given leases.

This is in order to help simulate the Infracloud configuration.

In order to do this, and hopefully get the CI test into a passing
state, I have increased the amount of memory allocated to each VM
as the realistic minimum for IPA is 1 GB, and introduced specific
stops and starts to VMs present during the test sequence in order
to limit the number of concurrent running virtual machines.

Change-Id: Ia612414edd80d154200e9df12b410a327db19cff