Allow a defined but not dhcp addressing_mode
In some work we're doing, we override addressing_mode with the
string 'static'. This should lead to getting network_info.json in our
configdrives, but it doesn't because the test logic in these tasks isn't
right. It has been tested, and the undefined test short circuits, so
we don't ever get errors whether the value is defined as dhcp, static,
or undefined.

Change-Id: I01f42c003dae21b3f25e2f760214ba876ae1aea9