Fix nics typing in configdrive generation
network_metadata's parameter "nics" is expected to be a list. In some
instances it will be left as a string unless the type is explicitly
set. Also the default of '' is incorrect since a string is not expected
but a list. Omitting it is the correct way, since it is not rquired.

Change-Id: If3376ee24507ed6f22e269e4440d2e371ff33eb4