Introduce longer deployment wait with multi-node testing
The nature of multi-host interactions and testing can
be extremely racey. Since IPA can offer SSH, the wait
step can be short circuited resulting in us not waiting
long enough.

As such, introducing a longer wait _and_ re-check of SSH
connectivity.

While this does not completely fix the underlying race, it
is intended to help the test be more reliable in CI.

Partial-Bug: #1559764
Change-Id: Ief9ff9b2d158fa5322b5bb9be74105ba3d13b7fe