Explicitly set language for DIB builds
Diskimage-builder passes the localization environment variables
through to the image being built. Ansible is explicitly setting
"en_US.UTF-8" by default. Depending on the DIB build underway,
and the base image, programs executing in that environment
may fail to execute if localization is not present in the base
image.

As such, we will override the settings for the DIB step.

Closes-Bug: #1563338
Change-Id: I05f0eeb71d3c227b2d8ece52c603c8cfe480c46c