Config drive spec uses network_data.json
The filename was incorrectly implemented as network_info.json instead
of network_data.json. Support for reading the correct file already
exists in glean. As such, we need to write both files for user
compatability in the short term, and remove the second file at a
later point in time.

A release note has been added detailing this as well.

Co-Authored-By: Julia Kreger <juliaashleykreger@gmail.com>
Change-Id: I4150f86fbc141dcd6fd12094acab1ca43d76910c