Fix DHCP test scenario
It appears that test nodes, once are explicitly powered off,
depending on libvirt, may completely vanish. As we only test
at the end of the process if we've observed additional DHCP
leases, we should move to only defining the VMs such that they
are powered on when the sequence is ready, not in advance of
the test sequence.

Moving to defining instead of creating does improve the overall
efficency of the sequence as no processes are spawning suddenly
using CPU and RAM during the start of the test. This also allows
us to remove and correct ansible virt module invocations.

The ansible virt module is an extra module, not a core module,
at least on Ubuntu Trusty, cannot have it's requirements installed
via pip.

Change-Id: I8948ba5a941384fdcacebcdfe9b6fd4fe8384cb1
Depends-On: I01885c397ed1e020aae306db1656678b26306532