Initialize 'auth' to empty dict on bifrost_configdrives_dynamic
The bifrost-configdrives-dynamic role invocation fails at
the update_facts_from_ironic include, due to initializing the
'auth' fact to None.
Ansible converts None to 'None', but then the openstack argument
spec expects to be a dict, which fails as it cannot convert it.
Setting this to an empty dict fixes the issue.

Change-Id: Ia807db74b6a2f0cc107216270f8538232019c607
Closes-Bug: LP1612586